+++ crates/web-sys/src/features/gen_GpuCanvasContext.rs
@@ -29,7 +29,7 @@ extern "C" {
     pub fn canvas(this: &GpuCanvasContext) -> ::js_sys::Object;
     #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "GpuCanvasConfiguration")]
-    # [wasm_bindgen (method , structural , js_class = "GPUCanvasContext" , js_name = configure)]
+    # [wasm_bindgen (catch , method , structural , js_class = "GPUCanvasContext" , js_name = configure)]
     #[doc = "The `configure()` method."]
     #[doc = ""]
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/GPUCanvasContext/configure)"]
@@ -38,10 +38,13 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
     #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
-    pub fn configure(this: &GpuCanvasContext, configuration: &GpuCanvasConfiguration);
+    pub fn configure(
+        this: &GpuCanvasContext,
+        configuration: &GpuCanvasConfiguration,
+    ) -> Result<(), JsValue>;
     #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "GpuTexture")]
-    # [wasm_bindgen (method , structural , js_class = "GPUCanvasContext" , js_name = getCurrentTexture)]
+    # [wasm_bindgen (catch , method , structural , js_class = "GPUCanvasContext" , js_name = getCurrentTexture)]
     #[doc = "The `getCurrentTexture()` method."]
     #[doc = ""]
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/GPUCanvasContext/getCurrentTexture)"]
@@ -50,7 +53,7 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
     #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
-    pub fn get_current_texture(this: &GpuCanvasContext) -> GpuTexture;
+    pub fn get_current_texture(this: &GpuCanvasContext) -> Result<GpuTexture, JsValue>;
     #[cfg(web_sys_unstable_apis)]
     # [wasm_bindgen (method , structural , js_class = "GPUCanvasContext" , js_name = unconfigure)]
     #[doc = "The `unconfigure()` method."]