+++ crates/web-sys/src/features/gen_GpuFeatureName.rs
@@ -14,6 +14,7 @@ pub enum GpuFeatureName {
     DepthClipControl = "depth-clip-control",
     Depth32floatStencil8 = "depth32float-stencil8",
     TextureCompressionBc = "texture-compression-bc",
+    TextureCompressionBcSliced3d = "texture-compression-bc-sliced-3d",
     TextureCompressionEtc2 = "texture-compression-etc2",
     TextureCompressionAstc = "texture-compression-astc",
     TimestampQuery = "timestamp-query",
@@ -22,4 +23,6 @@ pub enum GpuFeatureName {
     Rg11b10ufloatRenderable = "rg11b10ufloat-renderable",
     Bgra8unormStorage = "bgra8unorm-storage",
     Float32Filterable = "float32-filterable",
+    ClipDistances = "clip-distances",
+    DualSourceBlending = "dual-source-blending",
 }